+++ src/extra/react.rs
@@ -38,6 +38,7 @@ use std::marker::PhantomData;
 use bevy_ecs::{
     component::Component,
     entity::Entity,
+    query::With,
     schedule::Schedule,
     system::{Commands, Query},
 };
@@ -46,7 +47,7 @@ use bevy_render::view::visibility::Visibility;
 
 use crate::{
     access::{CurrentRef, NextRef},
-    pattern::StatePatternExtClone as _,
+    pattern::StatePattern as _,
     state::State,
 };
 
@@ -54,9 +55,11 @@ use crate::{
 ///
 /// Used in [`ReactPlugin<S>`].
 pub fn schedule_react<S: State + Eq>(schedule: &mut Schedule) {
-    schedule.add_systems(S::ANY.on_edge(
-        (despawn_on_exit::<S>, hide_on_exit::<S>),
-        show_on_enter::<S>,
+    schedule.add_systems((
+        S::ANY.on_disable((despawn_on_disable::<S>, hide_on_disable::<S>)),
+        S::ANY.on_enable(show_on_enable::<S>),
+        S::ANY.on_exit((despawn_on_exit::<S>, hide_on_exit::<S>)),
+        S::ANY.on_enter(show_on_enter::<S>),
     ));
 }
 
@@ -91,20 +94,51 @@ fn despawn_on_exit<S: State>(
     }
 }
 
+/// A component that sets the despawn behavior on any disable of the [`State`] type `S`.
+#[derive(Component, Default)]
+#[cfg_attr(feature = "bevy_reflect", derive(bevy_reflect::Reflect))]
+pub enum DespawnOnDisable<S: State> {
+    /// Despawn the entity on any exit.
+    Single,
+    #[default]
+    /// Despawn the entity and its descendants on any exit.
+    Recursive,
+    /// Despawn the entity's descendants on any exit.
+    Descendants,
+    #[doc(hidden)]
+    _PhantomData(PhantomData<S>),
+}
+
+fn despawn_on_disable<S: State>(
+    mut commands: Commands,
+    reaction_query: Query<(Entity, &DespawnOnDisable<S>)>,
+) {
+    for (entity, reaction) in &reaction_query {
+        match reaction {
+            DespawnOnDisable::Single => commands.entity(entity).despawn(),
+            DespawnOnDisable::Recursive => commands.entity(entity).despawn_recursive(),
+            DespawnOnDisable::Descendants => {
+                commands.entity(entity).despawn_descendants();
+            }
+            DespawnOnDisable::_PhantomData(_) => unreachable!(),
+        }
+    }
+}
+
 /// A component that shows an entity within a specific value of the [`State`] type `S`.
 ///
 /// - On enter, the visibility will be set to inherited.
 /// - On exit, the visibility will be set to hidden.
 #[derive(Component, Default)]
 #[cfg_attr(feature = "bevy_reflect", derive(bevy_reflect::Reflect))]
-pub struct StateVisible<S: State>(
+pub struct VisibleWhileIn<S: State>(
     /// The state during which the entity should be visible.
     pub S,
 );
 
 fn hide_on_exit<S: State + Eq>(
     state: CurrentRef<S>,
-    mut reaction_query: Query<(&mut Visibility, &StateVisible<S>)>,
+    mut reaction_query: Query<(&mut Visibility, &VisibleWhileIn<S>)>,
 ) {
     for (mut visibility, reaction) in &mut reaction_query {
         if state.is_in(&reaction.0) {
@@ -115,11 +149,35 @@ fn hide_on_exit<S: State + Eq>(
 
 fn show_on_enter<S: State + Eq>(
     state: NextRef<S>,
-    mut reaction_query: Query<(&mut Visibility, &StateVisible<S>)>,
+    mut reaction_query: Query<(&mut Visibility, &VisibleWhileIn<S>)>,
 ) {
     for (mut visibility, reaction) in &mut reaction_query {
         if state.will_be_in(&reaction.0) {
             *visibility = Visibility::Inherited;
         }
+    }
+}
+
+/// A component that shows an entity while the [`State`] type `S` is enabled.
+///
+/// - On any enable, the visibility will be set to inherited.
+/// - On any disable, the visibility will be set to hidden.
+#[derive(Component, Default)]
+#[cfg_attr(feature = "bevy_reflect", derive(bevy_reflect::Reflect))]
+pub struct VisibleWhileEnabled<S: State>(PhantomData<S>);
+
+fn hide_on_disable<S: State + Eq>(
+    mut reaction_query: Query<&mut Visibility, With<VisibleWhileEnabled<S>>>,
+) {
+    for mut visibility in &mut reaction_query {
+        *visibility = Visibility::Hidden;
+    }
+}
+
+fn show_on_enable<S: State + Eq>(
+    mut reaction_query: Query<&mut Visibility, With<VisibleWhileEnabled<S>>>,
+) {
+    for mut visibility in &mut reaction_query {
+        *visibility = Visibility::Inherited;
     }
 }